requirements.txt
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
bleach==4.1.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.2
fastapi==0.74.1
gunicorn==20.1.0
h11==0.13.0
httptools==0.3.0
idna==3.3
ipykernel==6.9.1
ipython==8.1.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
pystache==0.6.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pywin32==303
pywinpty==2.0.2
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
rsa==4.8
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
stack-data==0.2.0
starlette==0.17.1
terminado==0.13.1
testpath==0.6.0
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.4
watchgod==0.7
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.1
widgetsnbextension==3.5.2
